+++ NEWS.md
@@ -7,6 +7,7 @@
 
 #### Updates
 
+* `listen()` and `dial()` now return (invisible) zero rather than NULL upon success to better align with similar functions.
 * Allows setting the environment variable 'NANONEXT_ARM' prior to package installation
   + Fixes installation issues on certain ARM architectures
 * Deprecated functions 'send_vec' and 'recv_vec' removed.